move set_auto_role to it's own file
