+++ crack-core/src/commands/admin/mod.rs
@@ -21,6 +21,7 @@ pub mod print_settings;
 pub mod set_all_log_channel;
 pub mod set_all_log_channel_data;
 pub mod set_all_log_channel_old_data;
+pub mod set_auto_role;
 pub mod set_idle_timeout;
 pub mod set_prefix;
 pub mod set_welcome_settings;
@@ -54,10 +55,10 @@ pub use print_settings::*;
 pub use set_all_log_channel::*;
 pub use set_all_log_channel_data::*;
 pub use set_all_log_channel_old_data::*;
+pub use set_auto_role::*;
 pub use set_idle_timeout::*;
 pub use set_prefix::*;
-pub use set_welcome_settings::set_auto_role;
-pub use set_welcome_settings::set_welcome_settings;
+pub use set_welcome_settings::*;
 pub use unban::*;
 pub use unmute::*;
